Move utils to spark and yarn directories
